Small fixes: uninstall target guard, signed/unsigned compare, dead link
- CMakeLists.txt: guard the uninstall target with if(NOT TARGET ...).
  Redefining it breaks configuration when RtMidi is added as a subproject
  alongside another library that defines its own 'uninstall' target, which
  is fatal as of CMake 4.1. Fixes #369.

- RtMidi.cpp: CreateConnectedEndpointName compared 'int i' against
  'size_t nConnected'. Make i a size_t. Fixes #291.

- RtMidi.cpp: the mididev Google Groups thread referenced in WinMidiData is
  no longer publicly reachable. Replace the link with what it explained,
  keeping the attribution. Fixes #371.

+++ RtMidi.cpp
@@ -1399,7 +1399,7 @@ static CFStringRef CreateConnectedEndpointName( MIDIEndpointRef endpoint )
   CFMutableStringRef result = CFStringCreateMutable( NULL, 0 );
   CFStringRef str;
   OSStatus err;
-  int i;
+  size_t i;
 
   // Does the endpoint have connections?
   CFDataRef connections = NULL;
@@ -2687,7 +2687,7 @@ struct WinMidiData {
   DWORD lastTime;
   MidiInApi::MidiMessage message;
   std::vector<LPMIDIHDR> sysexBuffer;
-  CRITICAL_SECTION _mutex; // [Patrice] see https://groups.google.com/forum/#!topic/mididev/6OUjHutMpEo
+  CRITICAL_SECTION _mutex; // [Patrice] protects the sysex buffer requeue in midiInputCallback
 };
 
 //*********************************************************************//
